@@ -20,21 +20,21 @@
         }
     ],
     "require": {
-        "php": "^7.2|^8.0",
+        "php": "~8.1.0 || ~8.2.0 || ~8.3.0",
         "psr/log": "^1.0|^2.0|^3.0",
-        "firebase/php-jwt": "^3.0|^4.0|^5.0",
-        "psr/http-message": "^1.0",
-        "tuupola/http-factory": "^1.3",
-        "tuupola/callable-handler": "^1.0",
-        "psr/http-server-middleware": "^1.0"
+        "firebase/php-jwt": "^v6.8.0",
+        "psr/http-message": "^1.1 || ^2.0",
+        "tuupola/http-factory": "^1.4",
+        "tuupola/callable-handler": "^1.1",
+        "psr/http-server-middleware": "^1.0.2"
     },
     "require-dev": {
-        "squizlabs/php_codesniffer": "^3.7",
-        "overtrue/phplint": "^1.0",
-        "phpstan/phpstan": "^1.8",
-        "phpunit/phpunit": "^7.0|^8.5.30|^9.0",
+        "doctrine/coding-standard": "^12.0",
         "equip/dispatch": "^2.0",
-        "laminas/laminas-diactoros": "^2.0"
+        "infection/infection": "^0.27.2",
+        "laminas/laminas-diactoros": "^3.2",
+        "phpstan/phpstan": "^1.10.35",
+        "phpunit/phpunit": "^10.3.5"
     },
     "autoload": {
         "psr-4": {
@@ -43,12 +43,22 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Tuupola\\Middleware\\": "tests"
+            "Tuupola\\Tests\\Middleware\\": "tests"
         }
     },
     "extra": {
         "branch-alias": {
             "dev-3.x": "3.0.x-dev"
+        }
+    },
+    "config": {
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "infection/extension-installer": true
+        },
+        "platform": {
+          "php": "8.1.99"
         }
     }
 }
