@@ -29,123 +29,43 @@ SOFTWARE.
 /**
  * @see       https://github.com/tuupola/slim-jwt-auth
  * @see       https://appelsiini.net/projects/slim-jwt-auth
- * @license   https://www.opensource.org/licenses/mit-license.php
  */
 
 namespace Tuupola\Middleware;
 
-use Closure;
-use DomainException;
-use InvalidArgumentException;
-use Exception;
 use Firebase\JWT\JWT;
-use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use Psr\Log\LoggerInterface;
-use Psr\Log\LogLevel;
+use Psr\Log\NullLogger;
 use RuntimeException;
-use SplStack;
-use Tuupola\Middleware\DoublePassTrait;
-use Tuupola\Http\Factory\ResponseFactory;
-use Tuupola\Middleware\JwtAuthentication\RequestMethodRule;
-use Tuupola\Middleware\JwtAuthentication\RequestPathRule;
-use Tuupola\Middleware\JwtAuthentication\RuleInterface;
+use Throwable;
+
+use function array_key_exists;
+use function in_array;
+use function preg_match;
+use function sprintf;
+use function strtoupper;
 
 final class JwtAuthentication implements MiddlewareInterface
 {
-    use DoublePassTrait;
+    private readonly LoggerInterface $logger;
 
-    /**
-     * PSR-3 compliant logger.
-     * @var LoggerInterface|null
-     */
-    private $logger;
+    public function __construct(
+        private readonly JwtAuthenticationOption $options,
+        ?LoggerInterface $logger = null
+    ) {
+        $this->logger = $logger ?? new NullLogger();
+    }
 
-    /**
-     * Last error message.
-     * @var string
-     */
-    private $message;
-
-    /**
-     * The rules stack.
-     * @var SplStack<RuleInterface>
-     */
-    private $rules;
-
-    /**
-     * Stores all the options passed to the middleware.
-     *
-     * @var array{
-     *   secret?: string|array<string>,
-     *   secure: bool,
-     *   relaxed: array<string>,
-     *   algorithm: array<string>,
-     *   header: string,
-     *   regexp: string,
-     *   cookie: string,
-     *   attribute: string,
-     *   path: array<string>,
-     *   ignore: array<string>,
-     *   before: null|callable,
-     *   after: null|callable,
-     *   error: null|callable,
-     * }
-     */
-    private $options = [
-        "secure" => true,
-        "relaxed" => ["localhost", "127.0.0.1"],
-        "algorithm" => ["HS256", "HS512", "HS384"],
-        "header" => "Authorization",
-        "regexp" => "/Bearer\s+(.*)$/i",
-        "cookie" => "token",
-        "attribute" => "token",
-        "path" => ["/"],
-        "ignore" => [],
-        "before" => null,
-        "after" => null,
-        "error" => null
-    ];
-
-    /**
-     * @param array{
-     *   secret?: string|array<string>,
-     *   secure?: bool,
-     *   relaxed?: array<string>,
-     *   algorithm?: array<string>,
-     *   header?: string,
-     *   regexp?: string,
-     *   cookie?: string,
-     *   attribute?: string,
-     *   path?: array<string>,
-     *   ignore?: array<string>,
-     *   before?: null|callable,
-     *   after?: null|callable,
-     *   error?: null|callable,
-     * } $options
-     */
-    public function __construct(array $options = [])
-    {
-        /* Setup stack for rules */
-        $this->rules = new \SplStack;
-
-        /* Store passed in options overwriting any defaults. */
-        $this->hydrate($options);
-
-        /* If nothing was passed in options add default rules. */
-        /* This also means $options["rules"] overrides $options["path"] */
-        /* and $options["ignore"] */
-        if (!isset($options["rules"])) {
-            $this->rules->push(new RequestMethodRule([
-                "ignore" => ["OPTIONS"]
-            ]));
-            $this->rules->push(new RequestPathRule([
-                "path" => $this->options["path"],
-                "ignore" => $this->options["ignore"]
-            ]));
-        }
+    public function __invoke(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        callable $next
+    ): ResponseInterface {
+        return $this->process($request, new CallableHandler($next, $response));
     }
 
     /**
@@ -154,125 +74,39 @@ final class JwtAuthentication implements MiddlewareInterface
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         $scheme = $request->getUri()->getScheme();
-        $host = $request->getUri()->getHost();
+        $host   = $request->getUri()->getHost();
 
-        /* If rules say we should not authenticate call next and return. */
-        if (false === $this->shouldAuthenticate($request)) {
+        /* HTTP allowed only if secure is false or server is in relaxed array. */
+        if ($scheme !== 'https' && $this->options->secure === true && ! in_array($host, $this->options->relaxed)) {
+            throw new RuntimeException(sprintf(
+                'Insecure use of middleware over %s denied by configuration.',
+                strtoupper($scheme),
+            ));
+        }
+
+        try {
+            $token = $this->fetchToken($request);
+        } catch (TokenNotFound) {
             return $handler->handle($request);
         }
 
-        /* HTTP allowed only if secure is false or server is in relaxed array. */
-        if ("https" !== $scheme && true === $this->options["secure"]) {
-            if (!in_array($host, $this->options["relaxed"])) {
-                $message = sprintf(
-                    "Insecure use of middleware over %s denied by configuration.",
-                    strtoupper($scheme)
-                );
-                throw new RuntimeException($message);
-            }
-        }
-
-        /* If token cannot be found or decoded return with 401 Unauthorized. */
         try {
-            $token = $this->fetchToken($request);
-            $decoded = $this->decodeToken($token);
-        } catch (RuntimeException | DomainException $exception) {
-            $response = (new ResponseFactory)->createResponse(401);
-            return $this->processError($response, [
-                "message" => $exception->getMessage(),
-                "uri" => (string)$request->getUri()
-            ]);
+            $jwtDecodedToken = $this->decodeToken($token);
+        } catch (Throwable) {
+            return $handler->handle($request);
         }
-
-        $params = [
-            "decoded" => $decoded,
-            "token" => $token,
-        ];
 
         /* Add decoded token to request as attribute when requested. */
-        if ($this->options["attribute"]) {
-            $request = $request->withAttribute($this->options["attribute"], $decoded);
-        }
+        $request = $request->withAttribute($this->options->attribute, $jwtDecodedToken);
 
         /* Modify $request before calling next middleware. */
-        if (is_callable($this->options["before"])) {
-            $beforeRequest = $this->options["before"]($request, $params);
-            if ($beforeRequest instanceof ServerRequestInterface) {
-                $request = $beforeRequest;
-            }
-        }
+        $request = $this->options->before->__invoke($request, $jwtDecodedToken);
 
         /* Everything ok, call next middleware. */
         $response = $handler->handle($request);
 
         /* Modify $response before returning. */
-        if (is_callable($this->options["after"])) {
-            $afterResponse = $this->options["after"]($response, $params);
-            if ($afterResponse instanceof ResponseInterface) {
-                return $afterResponse;
-            }
-        }
-
-        return $response;
-    }
-
-    /**
-     * Set all rules in the stack.
-     *
-     * @param RuleInterface[] $rules
-     */
-    public function withRules(array $rules): self
-    {
-        $new = clone $this;
-        /* Clear the stack */
-        unset($new->rules);
-        $new->rules = new \SplStack;
-        /* Add the rules */
-        foreach ($rules as $callable) {
-            $new = $new->addRule($callable);
-        }
-        return $new;
-    }
-
-    /**
-     * Add a rule to the stack.
-     */
-    public function addRule(callable $callable): self
-    {
-        $new = clone $this;
-        $new->rules = clone $this->rules;
-        $new->rules->push($callable);
-        return $new;
-    }
-
-    /**
-     * Check if middleware should authenticate.
-     */
-    private function shouldAuthenticate(ServerRequestInterface $request): bool
-    {
-        /* If any of the rules in stack return false will not authenticate */
-        foreach ($this->rules as $callable) {
-            if (false === $callable($request)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Call the error handler if it exists.
-     *
-     * @param mixed[] $arguments
-     */
-    private function processError(ResponseInterface $response, array $arguments): ResponseInterface
-    {
-        if (is_callable($this->options["error"])) {
-            $handlerResponse = $this->options["error"]($response, $arguments);
-            if ($handlerResponse instanceof ResponseInterface) {
-                return $handlerResponse;
-            }
-        }
-        return $response;
+        return $this->options->after->__invoke($response, $jwtDecodedToken);
     }
 
     /**
@@ -281,11 +115,12 @@ final class JwtAuthentication implements MiddlewareInterface
     private function fetchToken(ServerRequestInterface $request): string
     {
         /* Check for token in header. */
-        $header = $request->getHeaderLine($this->options["header"]);
+        $header = $request->getHeaderLine($this->options->header);
 
-        if (false === empty($header)) {
-            if (preg_match($this->options["regexp"], $header, $matches)) {
-                $this->log(LogLevel::DEBUG, "Using token from request header");
+        if (empty($header) === false) {
+            if (preg_match($this->options->regexp, $header, $matches)) {
+                $this->logger->debug('Using token from request header');
+
                 return $matches[1];
             }
         }
@@ -293,222 +128,34 @@ final class JwtAuthentication implements MiddlewareInterface
         /* Token not found in header try a cookie. */
         $cookieParams = $request->getCookieParams();
 
-        if (isset($cookieParams[$this->options["cookie"]])) {
-            $this->log(LogLevel::DEBUG, "Using token from cookie");
-            if (preg_match($this->options["regexp"], $cookieParams[$this->options["cookie"]], $matches)) {
+        if (array_key_exists($this->options->cookie, $cookieParams)) {
+            $this->logger->debug('Using token from cookie');
+            if (preg_match($this->options->regexp, $cookieParams[$this->options->cookie], $matches)) {
                 return $matches[1];
             }
-            return $cookieParams[$this->options["cookie"]];
-        };
+
+            return $cookieParams[$this->options->cookie];
+        }
 
         /* If everything fails log and throw. */
-        $this->log(LogLevel::WARNING, "Token not found");
-        throw new RuntimeException("Token not found.");
+        $this->logger->debug('Token not found');
+
+        throw TokenNotFound::create();
     }
 
     /**
      * Decode the token.
-     *
-     * @return mixed[]
      */
-    private function decodeToken(string $token): array
+    private function decodeToken(string $token): JwtDecodedToken
     {
         try {
-            $decoded = JWT::decode(
-                $token,
-                $this->options["secret"],
-                (array) $this->options["algorithm"]
-            );
-            return (array) $decoded;
-        } catch (Exception $exception) {
-            $this->log(LogLevel::WARNING, $exception->getMessage(), [$token]);
+            $decoded = JWT::decode($token, $this->options->secret->__invoke($this->options->algorithm));
+        } catch (Throwable $exception) {
+            $this->logger->warning($exception->getMessage(), [$token]);
+
             throw $exception;
         }
-    }
 
-    /**
-     * Hydrate options from given array.
-     *
-     * @param mixed[] $data
-     */
-    private function hydrate(array $data = []): void
-    {
-        foreach ($data as $key => $value) {
-            /* https://github.com/facebook/hhvm/issues/6368 */
-            $key = str_replace(".", " ", $key);
-            $method = lcfirst(ucwords($key));
-            $method = str_replace(" ", "", $method);
-            if (method_exists($this, $method)) {
-                /* Try to use setter */
-                /** @phpstan-ignore-next-line */
-                call_user_func([$this, $method], $value);
-            } else {
-                /* Or fallback to setting option directly */
-                $this->options[$key] = $value;
-            }
-        }
-    }
-
-    /**
-     * Set path where middleware should bind to.
-     *
-     * @param string|string[] $path
-     */
-    private function path($path): void
-    {
-        $this->options["path"] = (array) $path;
-    }
-
-    /**
-     * Set path which middleware ignores.
-     *
-     * @param string|string[] $ignore
-     */
-    private function ignore($ignore): void
-    {
-        $this->options["ignore"] = (array) $ignore;
-    }
-
-    /**
-     * Set the cookie name where to search the token from.
-     */
-    private function cookie(string $cookie): void
-    {
-        $this->options["cookie"] = $cookie;
-    }
-
-    /**
-     * Set the secure flag.
-     */
-    private function secure(bool $secure): void
-    {
-        $this->options["secure"] = $secure;
-    }
-
-    /**
-     * Set hosts where secure rule is relaxed.
-     *
-     * @param string[] $relaxed
-     */
-    private function relaxed(array $relaxed): void
-    {
-        $this->options["relaxed"] = $relaxed;
-    }
-
-    /**
-     * Set the secret key.
-     *
-     * @param string|string[] $secret
-     */
-    private function secret($secret): void
-    {
-        if (false === is_array($secret) && false === is_string($secret) && ! $secret instanceof \ArrayAccess) {
-            throw new InvalidArgumentException(
-                'Secret must be either a string or an array of "kid" => "secret" pairs'
-            );
-        }
-        $this->options["secret"] = $secret;
-    }
-
-    /**
-     * Set the error handler.
-     */
-    private function error(callable $error): void
-    {
-        if ($error instanceof Closure) {
-            $this->options["error"] = $error->bindTo($this);
-        } else {
-            $this->options["error"] = $error;
-        }
-    }
-
-    /**
-     * Set the logger.
-     */
-    private function logger(LoggerInterface $logger = null): void
-    {
-        $this->logger = $logger;
-    }
-
-    /**
-     * Logs with an arbitrary level.
-     *
-     * @param mixed[] $context
-     */
-    private function log(string $level, string $message, array $context = []): void
-    {
-        if ($this->logger) {
-            $this->logger->log($level, $message, $context);
-        }
-    }
-
-    /**
-     * Set the attribute name used to attach decoded token to request.
-     */
-    private function attribute(string $attribute): void
-    {
-        $this->options["attribute"] = $attribute;
-    }
-
-    /**
-     * Set the header where token is searched from.
-     */
-    private function header(string $header): void
-    {
-        $this->options["header"] = $header;
-    }
-
-    /**
-     * Set the regexp used to extract token from header or environment.
-     */
-    private function regexp(string $regexp): void
-    {
-        $this->options["regexp"] = $regexp;
-    }
-
-    /**
-     * Set the allowed algorithms
-     *
-     * @param string|string[] $algorithm
-     */
-    private function algorithm($algorithm): void
-    {
-        $this->options["algorithm"] = (array) $algorithm;
-    }
-
-    /**
-     * Set the before handler.
-     */
-
-    private function before(callable $before): void
-    {
-        if ($before instanceof Closure) {
-            $this->options["before"] = $before->bindTo($this);
-        } else {
-            $this->options["before"] = $before;
-        }
-    }
-
-    /**
-     * Set the after handler.
-     */
-    private function after(callable $after): void
-    {
-        if ($after instanceof Closure) {
-            $this->options["after"] = $after->bindTo($this);
-        } else {
-            $this->options["after"] = $after;
-        }
-    }
-
-    /**
-     * Set the rules.
-     * @param RuleInterface[] $rules
-     */
-    private function rules(array $rules): void
-    {
-        foreach ($rules as $callable) {
-            $this->rules->push($callable);
-        }
+        return new JwtDecodedToken((array) $decoded, $token);
     }
 }
