@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
 
 Copyright (c) 2015-2022 Mika Tuupola
@@ -26,28 +28,36 @@ SOFTWARE.
 
 /**
  * @see       https://github.com/tuupola/slim-jwt-auth
- * @license   https://www.opensource.org/licenses/mit-license.php
+ * @see       https://appelsiini.net/projects/slim-jwt-auth
  */
 
-namespace Tuupola\Middleware;
+namespace Tuupola\Tests\Middleware\Assets;
 
-use Psr\Http\Message\ResponseInterface;
+use ArrayAccess;
 
-class TestAfterHandler
+/** @implements ArrayAccess<string, string> */
+class ArrayAccessImpl implements ArrayAccess
 {
-    public function __invoke(
-        ResponseInterface $response,
-        array $arguments
-    ) {
-        $response->getBody()->write(self::class);
-        return $response->withHeader("X-Brawndo", "plants crave");
+    /** @var array<string, string> */
+    private array $array = [];
+
+    public function offsetExists(mixed $offset): bool
+    {
+        return isset($this->array[$offset]);
     }
 
-    public static function after(
-        ResponseInterface $response,
-        array $arguments
-    ) {
-        $response->getBody()->write(self::class);
-        return $response->withHeader("X-Water", "like from toilet?");
+    public function offsetGet(mixed $offset): mixed
+    {
+        return $this->array[$offset];
+    }
+
+    public function offsetSet(mixed $offset, mixed $value): void
+    {
+        $this->array[$offset] = $value;
+    }
+
+    public function offsetUnset(mixed $offset): void
+    {
+        unset($this->array[$offset]);
     }
 }
